+++ sadu-queries2/src/test/java/de/chojo/sadu/queries/examples/ReadTest.java
@@ -9,10 +9,10 @@ package de.chojo.sadu.queries.examples;
 import de.chojo.sadu.PostgresDatabase;
 import de.chojo.sadu.mapper.PostgresqlMapper;
 import de.chojo.sadu.mapper.RowMapperRegistry;
+import de.chojo.sadu.queries.api.call.Call;
+import de.chojo.sadu.queries.api.call.calls.Calls;
 import de.chojo.sadu.queries.api.results.reading.Result;
 import de.chojo.sadu.queries.api.results.writing.ManipulationResult;
-import de.chojo.sadu.queries.call.Call;
-import de.chojo.sadu.queries.calls.Calls;
 import de.chojo.sadu.queries.configuration.QueryConfiguration;
 import de.chojo.sadu.queries.configuration.QueryConfigurationBuilder;
 import de.chojo.sadu.queries.examples.dao.User;
@@ -29,8 +29,8 @@ import java.util.Optional;
 import java.util.UUID;
 
 import static de.chojo.sadu.PostgresDatabase.createContainer;
-import static de.chojo.sadu.queries.call.Call.call;
-import static de.chojo.sadu.queries.call.adapter.impl.UUIDAdapter.AS_STRING;
+import static de.chojo.sadu.queries.api.call.Call.call;
+import static de.chojo.sadu.queries.call.adapter.UUIDAdapter.AS_STRING;
 
 @SuppressWarnings({"unused", "ResultOfMethodCallIgnored", "OptionalGetWithoutIsPresent", "RedundantExplicitVariableType"})
 public class ReadTest {
