+++ sadu-queries2/src/test/java/de/chojo/sadu/queries/examples/WriteTest.java
@@ -9,10 +9,10 @@ package de.chojo.sadu.queries.examples;
 import de.chojo.sadu.PostgresDatabase;
 import de.chojo.sadu.mapper.PostgresqlMapper;
 import de.chojo.sadu.mapper.RowMapperRegistry;
+import de.chojo.sadu.queries.api.call.Call;
 import de.chojo.sadu.queries.api.results.writing.ManipulationBatchResult;
 import de.chojo.sadu.queries.api.results.writing.ManipulationResult;
-import de.chojo.sadu.queries.call.Call;
-import de.chojo.sadu.queries.call.adapter.impl.UUIDAdapter;
+import de.chojo.sadu.queries.call.adapter.UUIDAdapter;
 import de.chojo.sadu.queries.configuration.QueryConfiguration;
 import de.chojo.sadu.queries.configuration.QueryConfigurationBuilder;
 import org.junit.jupiter.api.AfterEach;
@@ -26,7 +26,7 @@ import java.util.UUID;
 import java.util.stream.Stream;
 
 import static de.chojo.sadu.PostgresDatabase.createContainer;
-import static de.chojo.sadu.queries.call.adapter.impl.UUIDAdapter.AS_STRING;
+import static de.chojo.sadu.queries.call.adapter.UUIDAdapter.AS_STRING;
 
 @SuppressWarnings({"unused", "RedundantExplicitVariableType"})
 public class WriteTest {
